requirements.txt
python-dotenv==1.0.0
openai==0.27.7